api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
transformers==4.35.2
torch==2.1.0
torchvision==0.16.0
Pillow==10.0.1
boto3==1.34.0
numpy==1.24.3
requests==2.31.0
timm==0.9.12